+++ src/blogmore/generator.py
@@ -82,6 +82,7 @@ class SiteGenerator:
         output_dir: Path,
         site_title: str = "My Blog",
         site_subtitle: str = "",
+        site_description: str = "",
         site_url: str = "",
         posts_per_feed: int = 20,
         extra_stylesheets: list[str] | None = None,
@@ -101,6 +102,8 @@ class SiteGenerator:
             output_dir: Directory where generated site will be written
             site_title: Title of the blog site
             site_subtitle: Subtitle of the blog site
+            site_description: Default description used in metadata for pages that
+                              have no description of their own
             site_url: Base URL of the site
             posts_per_feed: Maximum number of posts to include in feeds (default: 20)
             extra_stylesheets: Optional list of URLs for additional stylesheets
@@ -116,6 +119,7 @@ class SiteGenerator:
         self.output_dir = output_dir
         self.site_title = site_title
         self.site_subtitle = site_subtitle
+        self.site_description = site_description
         self.site_url = normalize_site_url(site_url)
         self.posts_per_feed = posts_per_feed
         self.default_author = default_author
@@ -210,6 +214,7 @@ class SiteGenerator:
         context = {
             "site_title": self.site_title,
             "site_subtitle": self.site_subtitle,
+            "site_description": self.site_description,
             "site_url": self.site_url,
             "tag_dir": self.TAG_DIR,
             "category_dir": self.CATEGORY_DIR,